reorganize legend-application
