+++ .changeset/chilled-suns-burn.md
@@ -1,0 +1,8 @@
+---
+'@finos/legend-application': patch
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-graph': patch
+'@finos/legend-manual-tests': patch
+'@finos/legend-shared': patch
+'@finos/legend-studio': patch
+---
