+++ packages/legend-application/src/components/shared/LambdaParameterValuesEditor.tsx
@@ -22,8 +22,8 @@ import {
 } from '@finos/legend-graph';
 import { prettyCONSTName } from '@finos/legend-shared';
 import { observer } from 'mobx-react-lite';
-import type { LambdaParametersState } from '../stores/LambdaParameterState.js';
-import { useApplicationStore } from './ApplicationStoreProvider.js';
+import type { LambdaParametersState } from '../../stores/shared/LambdaParameterState.js';
+import { useApplicationStore } from '../ApplicationStoreProvider.js';
 import { BasicValueSpecificationEditor } from './BasicValueSpecificationEditor.js';
 
 // TODO we may want to further componentize this by removing the dialog wrapper
