+++ packages/legend-application/src/stores/shared/LambdaParameterState.ts
@@ -57,7 +57,7 @@ import {
   multiplicity_setUpperBound,
 } from './ValueSpecificationModifierHelper.js';
 import { format, addDays } from 'date-fns';
-import { DATE_FORMAT, DATE_TIME_FORMAT } from '../const.js';
+import { DATE_FORMAT, DATE_TIME_FORMAT } from '../../const.js';
 
 export enum PARAMETER_SUBMIT_ACTION {
   EXECUTE = 'EXECUTE',
