+++ packages/legend-application/style/components/shared/_documentation-link.scss
