+++ packages/legend-application/style/components/shared/_lambda-editor.scss
