+++ packages/legend-application/style/components/shared/_text-editor.scss
