+++ packages/legend-application/style/components/shared/_value-spec-editor.scss
